fmt and eliminate warning
