+++ src/lib.rs
@@ -6,17 +6,16 @@ extern crate alloc;
 extern crate volatile;
 #[macro_use]
 extern crate log;
-#[macro_use]
 extern crate aarch64;
 
 pub mod atags;
 pub mod consts;
+pub mod dma;
+pub mod emmc;
 pub mod gpio;
 pub mod interrupt;
 pub mod mailbox;
 pub mod mini_uart;
 pub mod pwm;
-pub mod dma;
 pub mod timer;
 pub mod util;
-pub mod emmc;
