+++ src/pwm.rs
@@ -1,18 +1,57 @@
 use crate::consts::*;
 use crate::timer::*;
-use crate::util::*;
-use alloc::vec::*;
-use core::mem::size_of;
-use core::ptr;
-use volatile::{ReadOnly, Volatile, WriteOnly};
+use volatile::Volatile;
 
+// GPIO clock (ref: peripherals page 107)
 const CLK_CTL_BUSY: u32 = (1 << 7);
 const CLK_CTL_KILL: u32 = (1 << 5);
 const CLK_CTL_ENAB: u32 = (1 << 4);
+pub const ARM_CM_BASE: usize = IO_BASE + 0x101000;
+pub const ARM_CM_GP0CTL: usize = ARM_CM_BASE + 0x70;
+pub const ARM_CM_GP0DIV: usize = ARM_CM_BASE + 0x74;
+pub const ARM_CM_PASSWD: u32 = (0x5A << 24);
 
+// PWM control register
+pub const ARM_PWM_CTL_PWEN1: usize = (1 << 0);
+pub const ARM_PWM_CTL_MODE1: usize = (1 << 1);
+pub const ARM_PWM_CTL_RPTL1: usize = (1 << 2);
+pub const ARM_PWM_CTL_SBIT1: usize = (1 << 3);
+pub const ARM_PWM_CTL_POLA1: usize = (1 << 4);
+pub const ARM_PWM_CTL_USEF1: usize = (1 << 5);
+pub const ARM_PWM_CTL_CLRF1: usize = (1 << 6);
+pub const ARM_PWM_CTL_MSEN1: usize = (1 << 7);
+pub const ARM_PWM_CTL_PWEN2: usize = (1 << 8);
+pub const ARM_PWM_CTL_MODE2: usize = (1 << 9);
+pub const ARM_PWM_CTL_RPTL2: usize = (1 << 10);
+pub const ARM_PWM_CTL_SBIT2: usize = (1 << 11);
+pub const ARM_PWM_CTL_POLA2: usize = (1 << 12);
+pub const ARM_PWM_CTL_USEF2: usize = (1 << 13);
+pub const ARM_PWM_CTL_MSEN2: usize = (1 << 15);
+
+// PWM status register
+pub const ARM_PWM_STA_FULL1: usize = (1 << 0);
+pub const ARM_PWM_STA_EMPT1: usize = (1 << 1);
+pub const ARM_PWM_STA_WERR1: usize = (1 << 2);
+pub const ARM_PWM_STA_RERR1: usize = (1 << 3);
+pub const ARM_PWM_STA_GAPO1: usize = (1 << 4);
+pub const ARM_PWM_STA_GAPO2: usize = (1 << 5);
+pub const ARM_PWM_STA_GAPO3: usize = (1 << 6);
+pub const ARM_PWM_STA_GAPO4: usize = (1 << 7);
+pub const ARM_PWM_STA_BERR: usize = (1 << 8);
+pub const ARM_PWM_STA_STA1: usize = (1 << 9);
+pub const ARM_PWM_STA_STA2: usize = (1 << 10);
+pub const ARM_PWM_STA_STA3: usize = (1 << 11);
+pub const ARM_PWM_STA_STA4: usize = (1 << 12);
+
+// PWM DMA configuration register
+pub const ARM_PWM_DMAC_DREQ__SHIFT: usize = 0;
+pub const ARM_PWM_DMAC_PANIC__SHIFT: usize = 8;
+pub const ARM_PWM_DMAC_ENAB: usize = (1 << 31);
+
+// PWM address map (ref: peripherals pages 141)
 #[repr(C)]
 #[allow(non_snake_case)]
-struct PWMAddressMap{
+struct PWMAddressMap {
     CTL: Volatile<u32>,
     STA: Volatile<u32>,
     DMAC: Volatile<u32>,
@@ -22,37 +61,34 @@ struct PWMAddressMap{
     FIF1: Volatile<u32>,
     reserve2: Volatile<u32>,
     RNG2: Volatile<u32>,
-    DAT2: Volatile<u32>
+    DAT2: Volatile<u32>,
 }
 
-
 pub struct PWMOutput {
-    pwm_address_map: &'static mut PWMAddressMap
+    pwm_address_map: &'static mut PWMAddressMap,
 }
 
 impl PWMOutput {
     pub fn new() -> PWMOutput {
         PWMOutput {
-            pwm_address_map: unsafe { &mut *(ARM_PWM_BASE as *mut PWMAddressMap) }
+            pwm_address_map: unsafe { &mut *(ARM_PWM_BASE as *mut PWMAddressMap) },
         }
     }
 
     pub fn start(&mut self, range: usize, use_fifo: bool) {
-        // warn!("range: {}", ((CLOCK_FREQ / CLOCK_DIVIDER + sample_rate / 2) / sample_rate));
         // init gpio clock
         let clk_clock = 6;
         let clk_source = 6 as u32;
         let n_div_i = 2 as u32;
         let n_div_f = 0 as u32;
         let n_div_mash = 0 as u32;
-        let gpio_clk_ctl: &'static mut Volatile<u32> = 
+        let gpio_clk_ctl: &'static mut Volatile<u32> =
             unsafe { &mut *((ARM_CM_GP0CTL + (clk_clock * 8)) as *mut Volatile<u32>) };
-        let gpio_clk_div: &'static mut Volatile<u32> = 
+        let gpio_clk_div: &'static mut Volatile<u32> =
             unsafe { &mut *((ARM_CM_GP0DIV + (clk_clock * 8)) as *mut Volatile<u32>) };
         // stop
         gpio_clk_ctl.write(ARM_CM_PASSWD | CLK_CTL_KILL);
         while gpio_clk_ctl.read() & CLK_CTL_BUSY != 0 {
-            //warn!("ctl: {}", gpio_clk_ctl.read());
             // do nothing
         }
         // start
@@ -62,36 +98,34 @@ impl PWMOutput {
         delay_us(10);
         gpio_clk_ctl.write(gpio_clk_ctl.read() | ARM_CM_PASSWD | CLK_CTL_ENAB);
         while gpio_clk_ctl.read() & CLK_CTL_BUSY == 0 {
-            //warn!("ctl: {}", gpio_clk_ctl.read());
             // do nothing
         }
 
-        //self.pwm_address_map.RNG1.write(((CLOCK_FREQ / CLOCK_DIVIDER + sample_rate / 2) / sample_rate) as u32);
-        //self.pwm_address_map.RNG2.write(((CLOCK_FREQ / CLOCK_DIVIDER + sample_rate / 2) / sample_rate) as u32);
         delay_us(2000);
         self.pwm_address_map.RNG1.write(range as u32);
         self.pwm_address_map.RNG2.write(range as u32);
-        self.pwm_address_map.CTL.write(self.pwm_address_map.CTL.read() |
-            ( ARM_PWM_CTL_PWEN1 | ARM_PWM_CTL_PWEN2) as u32);
-            //| ARM_PWM_CTL_MSEN1 | ARM_PWM_CTL_MSEN2) as u32);
+        self.pwm_address_map.CTL.write(
+            self.pwm_address_map.CTL.read() | (ARM_PWM_CTL_PWEN1 | ARM_PWM_CTL_PWEN2) as u32,
+        );
         self.pwm_address_map.STA.write(0);
         if use_fifo {
-            self.pwm_address_map.CTL.write(self.pwm_address_map.CTL.read() |
-                (ARM_PWM_CTL_USEF1 | ARM_PWM_CTL_USEF2 | ARM_PWM_CTL_CLRF1) as u32);
+            self.pwm_address_map.CTL.write(
+                self.pwm_address_map.CTL.read()
+                    | (ARM_PWM_CTL_USEF1 | ARM_PWM_CTL_USEF2 | ARM_PWM_CTL_CLRF1) as u32,
+            );
         }
-        delay_us(2000);
-        warn!("PWM CTL: {}", self.pwm_address_map.CTL.read());
         delay_us(2000);
     }
 
     pub fn dma_start(&mut self) {
-        self.pwm_address_map.DMAC.write((
-            ARM_PWM_DMAC_ENAB | (7 << ARM_PWM_DMAC_PANIC__SHIFT) | (7 << ARM_PWM_DMAC_DREQ__SHIFT)) as u32);
+        self.pwm_address_map.DMAC.write(
+            (ARM_PWM_DMAC_ENAB | (7 << ARM_PWM_DMAC_PANIC__SHIFT) | (7 << ARM_PWM_DMAC_DREQ__SHIFT))
+                as u32,
+        );
     }
 
     pub fn write(&mut self, channel: usize, data: u32) {
         if self.pwm_address_map.STA.read() & ARM_PWM_STA_BERR as u32 != 0 {
-            warn!("pwm write error!: {}", self.pwm_address_map.STA.read());
             self.pwm_address_map.STA.write(ARM_PWM_STA_BERR as u32);
         }
         if channel == 1 {
@@ -101,25 +135,20 @@ impl PWMOutput {
         }
     }
 
-    pub fn writeFIFO(&mut self, data: u32) {
+    pub fn write_fifo(&mut self, data: u32) {
         if self.pwm_address_map.STA.read() & ARM_PWM_STA_BERR as u32 != 0 {
-            // warn!("pwm beer error!: {}", self.pwm_address_map.STA.read());
             self.pwm_address_map.STA.write(ARM_PWM_STA_BERR as u32);
         }
         if self.pwm_address_map.STA.read() & ARM_PWM_STA_WERR1 as u32 != 0 {
-            // warn!("pwm write1 error!: {}", self.pwm_address_map.STA.read());
             self.pwm_address_map.STA.write(ARM_PWM_STA_WERR1 as u32);
         }
         if self.pwm_address_map.STA.read() & ARM_PWM_STA_GAPO1 as u32 != 0 {
-            // warn!("pwm gap1 error!: {}", self.pwm_address_map.STA.read());
             self.pwm_address_map.STA.write(ARM_PWM_STA_GAPO1 as u32);
         }
         if self.pwm_address_map.STA.read() & ARM_PWM_STA_GAPO2 as u32 != 0 {
-            // warn!("pwm gap2 error!: {}", self.pwm_address_map.STA.read());
             self.pwm_address_map.STA.write(ARM_PWM_STA_GAPO2 as u32);
         }
         while self.pwm_address_map.STA.read() & ARM_PWM_STA_FULL1 as u32 != 0 {
-            // warn!("pwm FIFO queue full! sta: {}", self.pwm_address_map.STA.read());
             // do nothing
         }
         self.pwm_address_map.FIF1.write(data);
